refactor: 중복되는 컴포넌트 추출과 커스텀 훅 작성
